backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==4.2.0",
    "django-cors-headers==4.0.0",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt>=5.5.0",
    "pyjwt==2.6.0",
    "python-dotenv==1.0.0",
    "requests==2.28.2",
]
